Pondering about how to effectively add threading...
Right now the elevator effectively teleports from one floor to the
desired floor. I'll need to add a time element or something or perhaps
a method that steps between floors. The problem being is that I'd then
need a way to "interrupt" this stepping between the floors to say to
pick up this person on <x> floor. Been awhile since I've made my own
threads in Java so will need to research for a bit.